Refactors to structure
