+++ src/main/client/public/index.html
@@ -44,6 +44,6 @@
 
     <!--    Cesium Script-->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/cesium/1.91.0/Cesium.js" integrity="sha512-iBK22GJMXjZVRsGPfjHIZfvIicCy12vfqX3fgC6AsZq/quNpnG/hnxve5i97fVfSyf7srMnVNtgojmc5pqMiyg==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
-
+<!--    <script src="https://cdnjs.cloudflare.com/ajax/libs/cesium/1.91.0/Cesium.min.js" integrity="sha512-ekYNnbOr96gMmw45tT6eFLWr9EIT4wgCOsmo85ZVAFsa7wWREj9I7PGbVgri4Sny6EFyJDlLfK+N9Z98z7YWBA==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>-->
   </body>
 </html>
