+++ src/main/java/com/ncrp/spring/app/controllers/ElasticSearchController.java
@@ -10,10 +10,8 @@ import org.elasticsearch.common.unit.DistanceUnit;
 import org.elasticsearch.index.query.QueryBuilders;
 import org.elasticsearch.search.builder.SearchSourceBuilder;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+
 import java.util.ArrayList;
 
 @RestController
@@ -32,6 +30,7 @@ public class ElasticSearchController
     }
 
     @GetMapping
+    @CrossOrigin(origins = "http://localhost:3000")
     /**
      * Get mapping for our endpoint.
      *
